requirements.txt

Flask-SQLAlchemy
psycopg2
psycopg2-binary
pydantic[email]
fastapi
Flask-Login
pytest
flask_testing
flask-cors
alembic
markdown
python-dotenv
openai
